+++ .prettierrc
@@ -1,0 +1,6 @@
+{
+  "tabWidth": 2,
+  "useTabs": false,
+  "organizeImportsSkipDestructiveCodeActions": true,
+  "singleQuote": false
+}
