+++ index.mjs
@@ -5,8 +5,8 @@ import webpack from "webpack";
 import webpackDevMiddleware from "webpack-dev-middleware";
 import webpackHotMiddleware from "webpack-hot-middleware";
 import MainRouter from "./routes/main.routes.mjs";
-import UserRouter from "./routes/user.routes.mjs";
 import SongsRouter from "./routes/songs.routes.mjs";
+import UserRouter from "./routes/user.routes.mjs";
 import webpackConfig from "./webpack_conf/webpack.dev.js";
 
 // Initialise Express instance
