+++ models/index.model.mjs
@@ -2,8 +2,8 @@ import { Sequelize } from "sequelize";
 import url from "url";
 import allConfig from "../config/config.js";
 
-import usersModel from "./users.mjs";
-import songsModel from "./songs.mjs";
+import songsModel from "./songs.model.mjs";
+import usersModel from "./users.model.mjs";
 
 const env = process.env.NODE_ENV || "development";
 const config = allConfig[env];
