--- models/songs.mjs
@@ -1,27 +1,0 @@
-export default function songsModel(sequelize, DataTypes) {
-  return sequelize.define('songs', {
-    id: {
-      allowNull: false,
-      autoIncrement: true,
-      primaryKey: true,
-      type: DataTypes.INTEGER,
-    },
-    title: {
-      type: DataTypes.STRING,
-    },
-    songData: {
-      type: DataTypes.STRING,
-    },
-    creatorId: {
-      type: DataTypes.STRING,
-    },
-    createdAt: {
-      allowNull: false,
-      type: DataTypes.DATE,
-    },
-    updatedAt: {
-      allowNull: false,
-      type: DataTypes.DATE,
-    },
-  }, { underscored: true });
-}
