+++ models/songs.model.mjs
@@ -1,0 +1,31 @@
+export default function songsModel(sequelize, DataTypes) {
+  return sequelize.define(
+    "songs",
+    {
+      id: {
+        allowNull: false,
+        autoIncrement: true,
+        primaryKey: true,
+        type: DataTypes.INTEGER,
+      },
+      title: {
+        type: DataTypes.STRING,
+      },
+      songData: {
+        type: DataTypes.STRING,
+      },
+      creatorId: {
+        type: DataTypes.STRING,
+      },
+      createdAt: {
+        allowNull: false,
+        type: DataTypes.DATE,
+      },
+      updatedAt: {
+        allowNull: false,
+        type: DataTypes.DATE,
+      },
+    },
+    { underscored: true }
+  );
+}
