+++ package.json
@@ -39,6 +39,7 @@
     "html-webpack-plugin": "^4.5.1",
     "mini-css-extract-plugin": "^1.3.3",
     "nodemon": "^2.0.7",
+    "prettier-plugin-organize-imports": "^3.1.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-refresh": "^0.9.0",
