+++ src/App.jsx
@@ -1,6 +1,6 @@
-import React, { useState } from "react";
-import { Routes, Route } from "react-router-dom";
-import Keeb from "./components/keyboard.jsx";
+import React from "react";
+import { Route, Routes } from "react-router-dom";
+import Keeb from "./components/keyboard/Keyboard.jsx";
 import Login from "./components/login/Login.jsx";
 import Root from "./components/Root.jsx";
