+++ src/components/keyboard/Keyboard.jsx
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { useRef, useState } from "react";
+import React, { useRef, useState } from "react";
 import Keyboard from "react-simple-keyboard";
 // import JZZ from 'jazz-midi'
